+++ algo/atbash.py
@@ -1,4 +1,4 @@
-from ciphers import ciphers
+from base.ciphers import ciphers
 class atbash(ciphers):
     def __init__(self):
         super().__init__()
