+++ algo/base/alphabet.py
